Address PR comments:
- move helper vals and funs from toplevel to abstract superclass
- restore accidentally deleted V1toV2 test
